policyrep: move filename transitions to avtab
Due to the move of filename transitions to the avtab transition
structure in libsepol, setools also need to be updated to access these
new structures.

This patch adds a linked list of filename transitions to the transition
structure added in the previous patch. It also updates FileNameTERule to
accept new filename transition structure and updates TERuleIterator to
include inner filename transitions in the iteration.

Signed-off-by: Juraj Marcin <juraj@jurajmarcin.com>