policyrep: add prefix/suffix filename transitions support to avtab
Currently, filename type transitions support only exact name matching.
However, in practice, the names contain variable parts. This leads to
many duplicated rules in the policy that differ only in the part of the
name, or it is even impossible to cover all possible combinations.

This patch implements required changes to structures same as in the
libsepol and other internal structures to support prefix/suffix filename
transitions. On top of that, it extends the FileNameTERule and
TERuleIterator to include these new rules.

Signed-off-by: Juraj Marcin <juraj@jurajmarcin.com>